// models/json.model.ts
export class Aluno{
    key?:string;
    nome: string;
    matricula: string;
    curso: string;
    uid: string;
}
export class Pergunta{
    key?:string;
    assunto: string;
    pergunta: string;
    username: string;
    useremail: string;
    uid: string;
    data: string;
    resposta: string;
}